UI: removed old cards from homescreen in favor for extracted cards from browser
