+++ App/src/desktopMain/kotlin/dev/dubsky/aiko/screens/BrowseScreen.kt
@@ -33,6 +33,7 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.unit.dp
+import dev.dubsky.aiko.components.card.AnimeCard
 import dev.dubsky.aiko.config.ConfigManager
 import dev.dubsky.aiko.data.Anime
 import dev.dubsky.aiko.graphql.type.MediaSeason
@@ -282,7 +283,7 @@ fun BrowseScreen(
                 horizontalArrangement = Arrangement.spacedBy(8.dp)
             ) {
                 items(displayedAnime) { anime ->
-                    AnimeCardO(
+                    AnimeCard(
                         anime = anime,
                         onClick = { onAnimeSelected(anime) },
                         modifier = Modifier
@@ -352,96 +353,7 @@ private fun <T> FilterDropdown(
     }
 }
 
-@Composable
-fun AnimeCardO(
-    anime: Anime,
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Card(
-        onClick = onClick,
-        modifier = modifier,
-        shape = RoundedCornerShape(12.dp),
-        elevation = CardDefaults.cardElevation(4.dp)
-    ) {
-        Box(modifier = Modifier.fillMaxSize()) {
-            AsyncImage(
-                model = anime.coverImage,
-                contentDescription = anime.title,
-                contentScale = ContentScale.Crop,
-                modifier = Modifier.fillMaxSize()
-            )
 
-            Box(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .background(
-                        Brush.verticalGradient(
-                            colors = listOf(
-                                Color.Transparent,
-                                Color.Black.copy(alpha = 0.7f)
-                            ),
-                            startY = 0.6f
-                        )
-                    )
-            )
-
-            Column(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .align(Alignment.BottomStart)
-                    .padding(12.dp)
-            ) {
-                Text(
-                    text = anime.title,
-                    style = MaterialTheme.typography.titleMedium,
-                    color = Color.White,
-                    maxLines = 2,
-                    overflow = TextOverflow.Ellipsis,
-                    modifier = Modifier.padding(bottom = 4.dp)
-                )
-
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    modifier = Modifier.padding(bottom = 4.dp)
-                ) {
-                    if (anime.season != MediaSeason.UNKNOWN__) {
-                        Text(
-                            text = anime.season.name.capitalize(),
-                            style = MaterialTheme.typography.labelSmall,
-                            color = Color.White.copy(alpha = 0.8f),
-                            modifier = Modifier.padding(end = 8.dp)
-                        )
-                    }
-                    if (anime.seasonYear > 0) {
-                        Text(
-                            text = anime.seasonYear.toString(),
-                            style = MaterialTheme.typography.labelSmall,
-                            color = Color.White.copy(alpha = 0.8f)
-                        )
-                    }
-                }
-
-                Row(
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Icon(
-                        imageVector = Icons.Filled.Star,
-                        contentDescription = null,
-                        tint = Color(0xFFFFD700),
-                        modifier = Modifier.size(16.dp)
-                    )
-                    Text(
-                        text = "%.1f".format(anime.rating / 10f),
-                        style = MaterialTheme.typography.labelSmall,
-                        color = Color.White,
-                        modifier = Modifier.padding(start = 4.dp)
-                    )
-                }
-            }
-        }
-    }
-}
 
 @Composable
 fun RatingBar(rating: Float) {
